--- a/Tests/NIOHTTP1Tests/HTTPDecoderTest+XCTest.swift
+++ b/Tests/NIOHTTP1Tests/HTTPDecoderTest+XCTest.swift
@@ -37,6 +37,8 @@ extension HTTPDecoderTest {
                 ("testCorrectlyMaintainIndicesWhenDiscardReadBytes", testCorrectlyMaintainIndicesWhenDiscardReadBytes),
                 ("testDropExtraBytes", testDropExtraBytes),
                 ("testDontDropExtraBytes", testDontDropExtraBytes),
+                ("testExtraCRLF", testExtraCRLF),
+                ("testSOURCEDoesntExplodeUs", testSOURCEDoesntExplodeUs),
            ]
    }
 }