@Lukasa 2018-06-05
The `http_parser` folks claimed a 15% performance improvement in their benchmark. Naturally we see less of that because our bench has way more code in it, but here's some perf. All of this is without pipelining.

Before:

```
wrk -c 400 -d 1m -t 2 --latency http://localhost:8888
Running 1m test @ http://localhost:8888
  2 threads and 400 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    19.19ms   10.43ms 559.46ms   96.01%
    Req/Sec    10.52k     0.86k   12.67k    76.92%
  Latency Distribution
     50%   18.00ms
     75%   20.77ms
     90%   24.12ms
     99%   37.87ms
  1256058 requests in 1.00m, 61.09MB read
  Socket errors: connect 0, read 276, write 2, timeout 0
Requests/sec:  20928.74
Transfer/sec:      1.02MB
```

After:

```
wrk -c 400 -d 1m -t 2 --latency http://localhost:8888
Running 1m test @ http://localhost:8888
  2 threads and 400 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    18.33ms    9.07ms 505.86ms   95.24%
    Req/Sec    10.99k   718.31    13.53k    75.67%
  Latency Distribution
     50%   17.35ms
     75%   19.91ms
     90%   22.91ms
     99%   33.98ms
  1312015 requests in 1.00m, 63.81MB read
  Socket errors: connect 0, read 269, write 3, timeout 0
Requests/sec:  21861.03
Transfer/sec:      1.06MB
```

Small improvement there, about 4.5%.