@Lukasa 2018-06-05
All the latencies have improved too, just seems like a net win for us.